better clause strengthening
Original minisat code seems incorrect. Consider a clause
  c0 c1 c2 c3

Strengthening removes a literal. Say it removed c1. The new clause is
  c0 c2 c3

if c2 happens to be l_False in the current context and c3 is not, then,
the clause should be
   c0 c3 c2
because c3 should be watched instead of c2.

Furthermore, if any of the c_i are l_True, then the clause is satisfied
and can be removed.