part logging for ccmin=2
With ccmin=2, a literal 'p' in a learned clause is resolved with clauses
whose literals are already in the learned clause.

That is, if the learned clause 'c' has a literal 'l', and there are a
bunch of clauses C such that C -> ~l, and Var(C) \subseteq Var(c), then
the learned clause 'c' is replaced by
   c   C
---------- ~l
    c\{~l}

The partition code computes the join of all partitions of clauses in C
and adds them to the partition of 'c'.