fix: handling of duplicate clauses
strengthening done by the simplifier might create a clause already in
the database. For proof-traversal, we silently remove them.

This requires to never deleting core clauses during replay. But that is
probably a good idea anyhow.

Not deleting clauses can lead to a situation where the database contains
two identical clauses C1 and C2. At some point the solver removed C1 and
used C2. During replay, C2 is a duplicate and is removed. But C1 should
never be removed.