fix: in replay ignore satisfied clauses
SimpSolver seems to add a lot of redundant clauses. With it, it is
possible that a clause is in the proof (i.e., marked core), but is
already satisfied when is about to be added in replay. A typical case is
when another identical clause is already in the database and is a reason
for some variable.