NEW: ordered_propagate mode
in ordered_propagate mode the propagation is done in the partition
order. Smallest to largest. That is, first clauses in partition 1 are
propagated, then in partitions 1 and 2, then 1, 2, and 3. etc.

This is the naive implementation that simply skips clauses that are not
currently being propagated.

A smarter implementation is to sort the watch lists, and maintain them
in sorted order during propagation. That is, in ordered_mode,
uncheckedEnqueue should add clauses to proper places in the watch list.