fix: core clause marking
Consider the following case:

G = !b , !c , !d, (a | d)

and a lemma L = (a | b | c) that we are validating. That is, we want to
show that G -> (a | b | c)

the trail is
!b, !c, !d

The existing code would add !a to the trail (because b and c are already
false) and infer that !d and (a | d) are in the core.

However, the correct behavior is to also mark !b and !c as part of the core.